Update to do reconfigure, when the main db is first created.
:Release Notes:
Update to do reconfigure, when the main db is first created.

:Detailed Notes:
If layers.json is simple and there is no selectionChanged event,
reconfigure is not performed even if the main db is first created.
As a result, post processing is not performed.
So update to do reconfigure at least once.

:Testing Performed:
Local test : OK

:QA Notes:

:Issues Addressed:
[PLAT-114694] Provide blocked app & service list using configd's post process.

Change-Id: I7ae1391906cc0b06c8c1fe504e14f7310bb8272d